requirements.txt
requests>=2.32.0
python-libsbml>=5.20.0
graphviz>=0.21
matplotlib>=3.10.0
numpy>=2.0.0
networkx>=3.6.0
scipy>=1.10.0
